YHErasymchuk - models

+++ database/migrations/2019_08_31_131849_create_cities_table.php
@@ -15,7 +15,7 @@ class CreateCitiesTable extends Migration
     {
         Schema::create('cities', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('country_id');
+            $table->unsignedBigInteger('country_id');
             $table->string('name');
             $table->timestamps();
         });
